refactor: consolidate industry service layers and compatibility wrappers
