+++ .gitignore
@@ -61,3 +61,4 @@ NOTES.local.md
 .github/copilot-instructions.md
 .github/instructions/indy_hub.instructions.md
 docs/manual-prod-validation-checklist.md
+graphify-out/
