+++ indy_hub/tasks/material_exchange_contracts.py
@@ -33,7 +33,6 @@ from indy_hub.models import (
     MaterialExchangeConfig,
     MaterialExchangeSellOrder,
     MaterialExchangeSettings,
-    MaterialExchangeStock,
     NotificationWebhook,
     NotificationWebhookMessage,
 )
@@ -52,25 +51,24 @@ from indy_hub.services.esi_client import (
     get_retry_after_seconds,
     shared_client,
 )
+from indy_hub.services.material_exchange_contract_helpers import (
+    get_config_location_ids,
+    get_config_location_match_names,
+    get_config_location_summary,
+    get_location_name,
+    is_transient_esi_error,
+    log_buy_order_transactions,
+    log_sell_order_transactions,
+    normalize_esi_mapping,
+    normalize_location_match_name,
+)
 from indy_hub.utils.analytics import emit_analytics_event
 from indy_hub.utils.db_retry import update_or_create_with_mysql_retry
-from indy_hub.utils.eve import PLACEHOLDER_PREFIX, get_type_name, resolve_location_name
-from indy_hub.utils.material_exchange_transactions import (
-    upsert_material_exchange_transaction,
-)
+from indy_hub.utils.eve import get_type_name
 
 logger = get_extension_logger(__name__)
 
-# Cache for structure names to avoid repeated lookups within a task run.
-_structure_name_cache: dict[int, str | None] = {}
-
-
-def _is_transient_esi_error(exc: ESIClientError) -> bool:
-    try:
-        status_code = int(exc.status_code or 0)
-    except (TypeError, ValueError):
-        status_code = 0
-    return 500 <= status_code < 600
+_is_transient_esi_error = is_transient_esi_error
 
 
 def _log_contract_cache_status_for_validation_skip(corporation_id: int) -> None:
@@ -95,179 +93,16 @@ def _log_contract_cache_status_for_validation_skip(corporation_id: int) -> None:
 
 
 def _normalize_esi_mapping(payload, *, context: str) -> dict | None:
-    """Return a dict from an ESI payload or None if unsupported."""
-    if isinstance(payload, dict):
-        return payload
-    for attr in ("model_dump", "dict", "to_dict"):
-        converter = getattr(payload, attr, None)
-        if callable(converter):
-            try:
-                result = converter()
-            except Exception:  # pragma: no cover - defensive
-                result = None
-            if isinstance(result, dict):
-                return result
-    logger.warning(
-        "Unexpected %s payload type for material exchange contracts: %s",
-        context,
-        type(payload).__name__,
-    )
-    return None
+    return normalize_esi_mapping(payload, context=context, logger=logger)
 
 
-def _log_sell_order_transactions(order: MaterialExchangeSellOrder) -> None:
-    _transaction, created = upsert_material_exchange_transaction(order)
-    if not created:
-        return
-
-    for item in order.items.all():
-        stock_item, _created = MaterialExchangeStock.objects.get_or_create(
-            config=order.config,
-            type_id=item.type_id,
-            defaults={"type_name": item.type_name},
-        )
-        stock_item.quantity += item.quantity
-        stock_item.save()
-
-
-def _log_buy_order_transactions(order: MaterialExchangeBuyOrder) -> None:
-    _transaction, created = upsert_material_exchange_transaction(order)
-    if not created:
-        return
-
-    for item in order.items.all():
-        try:
-            stock_item = order.config.stock_items.get(type_id=item.type_id)
-            stock_item.quantity = max(stock_item.quantity - item.quantity, 0)
-            stock_item.save()
-        except MaterialExchangeStock.DoesNotExist:
-            continue
-
-
-def _is_placeholder_location_name(name: str | None) -> bool:
-    return not name or str(name).startswith(PLACEHOLDER_PREFIX)
-
-
-def _normalize_location_match_name(name: str | None) -> str | None:
-    if _is_placeholder_location_name(name):
-        return None
-
-    normalized = str(name).strip()
-    if not normalized:
-        return None
-
-    if " > " in normalized:
-        normalized = normalized.split(" > ", 1)[0].strip()
-
-    return normalized.casefold() or None
-
-
-def _get_location_name(location_id: int | None) -> str | None:
-    """Resolve a contract location name using the shared cache-aware resolver."""
-
-    if not location_id:
-        return None
-
-    try:
-        normalized_location_id = int(location_id)
-    except (TypeError, ValueError):
-        return None
-
-    if normalized_location_id in _structure_name_cache:
-        return _structure_name_cache[normalized_location_id]
-
-    name: str | None = None
-
-    try:
-        direct_name = resolve_location_name(
-            normalized_location_id,
-            force_refresh=False,
-            allow_public=True,
-        )
-        if not _is_placeholder_location_name(direct_name):
-            name = str(direct_name)
-    except Exception:
-        logger.debug(
-            "Shared location resolver failed for %s",
-            normalized_location_id,
-            exc_info=True,
-        )
-
-    _structure_name_cache[normalized_location_id] = name
-    return name
-
-
-def _get_config_locations(config) -> list[dict[str, int | str]]:
-    rows: list[dict[str, int | str]] = []
-    try:
-        for location in config.accepted_locations.all().order_by("sort_order", "id"):
-            rows.append(
-                {
-                    "structure_id": int(location.structure_id),
-                    "structure_name": str(location.structure_name or ""),
-                    "hangar_division": int(location.hangar_division),
-                }
-            )
-    except Exception:
-        rows = []
-
-    if rows:
-        return rows
-
-    structure_id = getattr(config, "structure_id", None)
-    hangar_division = getattr(config, "hangar_division", None)
-    if not structure_id or not hangar_division:
-        return []
-
-    return [
-        {
-            "structure_id": int(structure_id),
-            "structure_name": str(getattr(config, "structure_name", "") or ""),
-            "hangar_division": int(hangar_division),
-        }
-    ]
-
-
-def _get_config_location_ids(config) -> set[int]:
-    location_ids: set[int] = set()
-    for location in _get_config_locations(config):
-        try:
-            structure_id = int(location.get("structure_id") or 0)
-        except (TypeError, ValueError):
-            structure_id = 0
-        if structure_id > 0:
-            location_ids.add(structure_id)
-    return location_ids
-
-
-def _get_config_location_summary(config) -> str:
-    labels: list[str] = []
-    for location in _get_config_locations(config):
-        structure_id = int(location.get("structure_id") or 0)
-        structure_name = str(location.get("structure_name") or "").strip()
-        hangar_division = int(location.get("hangar_division") or 0)
-        label = structure_name or f"Structure {structure_id}"
-        if hangar_division > 0:
-            label = f"{label} / Hangar {hangar_division}"
-        labels.append(label)
-    return ", ".join(label for label in labels if label)
-
-
-def _get_config_location_match_names(config) -> set[str]:
-    names: set[str] = set()
-
-    for location in _get_config_locations(config):
-        configured_name = _normalize_location_match_name(location.get("structure_name"))
-        if configured_name:
-            names.add(configured_name)
-
-        resolved_name = _normalize_location_match_name(
-            _get_location_name(location.get("structure_id"))
-        )
-        if resolved_name:
-            names.add(resolved_name)
-
-    return names
+_log_sell_order_transactions = log_sell_order_transactions
+_log_buy_order_transactions = log_buy_order_transactions
+_get_config_location_ids = get_config_location_ids
+_get_config_location_summary = get_config_location_summary
+_get_config_location_match_names = get_config_location_match_names
+_normalize_location_match_name = normalize_location_match_name
+_get_location_name = get_location_name
 
 
 @shared_task(
